Everywhere: Remove ILoader
This was a cool idea, but realistically there is no reason for CT to
support multiple languages. If we did want to add another language,
this abstraction would be 20-30% of the total work required to stuff
another engine into this mod. It adds unnecessary complication for no
benefit, so it's finally time to get rid of it.