Renderer: Accept Long as color instead of Int
This reverts a change I made a few weeks ago. Color is representible in
32 bits, and this is how MC stores it, however when Rhino is converting
a number to an Int, if it is larger than Int.MAX_VALUE it will error.
This means that if we use Int, it is impossible for a user to pass in a
color like 0xFFFF_FFFF.