Mixins: Fix redirect() not considering command of target node
Consider the following situation:

    val node = dispatcher.register(literal("foo")
    	.then(argument("arg", StringArgumentType.string())
    		.executes {
    			ChatLib.chat("foo with arg")
    			1
    		})
    	.executes {
    		ChatLib.chat("foo without arg")
    		1
    	})

    dispatcher.register(literal("bar").redirect(node))

With this setup, "/bar test" will work as expected, however "/bar" will
fail! Redirect nodes currently do not consider the command of the node
they redirect to, only their children.

As the comment is this Mixin indicates, this is more of a band-aid fix
than a proper one since any redirect modifiers used in the redirect()
call will be ignored if there are no arguments following the
redirection. This isn't so important for the purposes of the CT-provided
API, but may affect commands from other mods. We'll have to keep an eye
on this and address it if it becomes a problem.

Relevant issue: Mojang/brigadier#46