Add a KSP project to generate a TypeScript .d.ts file
This generated file contains declarations corresponding to the CT types,
as well as any used MC types. There are a few errors with the generated
file, but it is good enough to provide good code completion in 99.9% of
all cases. Fixing the remaining issues will require a lot of work, so it
is good enough for now.