KeyBind: Do not allow the use of uninitialized KeyBind object
In Kotlin, the initialization order of classes is:
  - primary constructor, if applicable
  - field intialization and init blocks, in the order they appear
  - the secondary constructor that was invoked, if applicable

KeyBind has no primary constructor, but it does have an init block and
secondary constructors. Inside the init block, we added the keybind to
a static list that is used elsewhere, primarily for initialization.
However, when this init block runs, the KeyBind object is not fully
initialized; specifically, the keyBinding member has not been assigned
(it is assigned in the secondary constructors).

If the code that used this list ran after the KeyBind was added to the
list but before the constructor finished running, it would lead to an
NPE. This is of course very volatile, as it depends on precise timing
between multiple threads.