Image: Allow for null textures
When the Image is initially created, it has to wait one more tick to create the Texture. This would cause problems if the user calls Renderer.bindTexture before that happens, resulting in a crash. We solve this by just providing 0 as the id, which will just render black for that one tick before it gets loaded.